// frontend/homework/11/quote-app/src/types/quote.tsx
// Generated by https://quicktype.io

import { ITag } from "./tag";

export interface ApiQuote {
    _id:          string;
    content:      string;
    author:       string;
    tags:         string[];
    authorSlug:   string;
    length:       number;
    dateAdded:    string;
    dateModified: string;
}

export interface IQuote {
    _id: string;
    content: string;
    author: string;
    tags: string[];
    dateAdded: string;
}

export interface IFetchQuote{
    fetchQuotesApi: (numQuotes: number, setLoading: React.Dispatch<React.SetStateAction<boolean>>) => Promise<any>;
}